requirements.txt
asgiref==3.8.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
diskcache==5.6.3
filelock==3.13.3
Flask==3.0.3
fsspec==2024.3.1
huggingface-hub==0.22.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
llama_cpp_python==0.2.61
MarkupSafe==2.1.5
munch==4.0.0
numpy==1.26.4
packaging==24.0
PyYAML==6.0.1
requests==2.31.0
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
